Move init under setup category
